Bug fix to support host only ssh.
Basically if user is not specified it would execute ssh like

ssh -lundefined foo.bar.com ls

Which does not work. In ssh user is optional and if current user has
access he will automagically be sshed in.
In my case i tend to have this configuration in .ssh/config so i almost never
specify a username